no longer store all of file in memory
