+++ hashmap.h
@@ -1,5 +1,4 @@
 #include "array_list.h"
-#include "read_file.h"
 
 #ifndef HASHMAP_H
 #define HASHMAP_H
@@ -11,7 +10,8 @@ typedef struct
     ArrayList *items[HASHMAP_SIZE];
 } Hashmap;
 
-Hashmap *hashmap_from_words(char words[MAX_WORD_COUNT][MAX_WORD_LENGTH], size_t word_count);
+Hashmap *hashmap_new();
+Hashmap *hashmap_from_file(char *path);
 void hashmap_free(Hashmap *map);
 size_t *hashmap_get(Hashmap *map, char *word);
 bool hashmap_insert(Hashmap *map, char word[ARRAY_LIST_WORD_MAX_LENGTH]);
