+++ words.h
@@ -1,0 +1,6 @@
+#ifndef READFILE_H
+#define READFILE_H
+
+char *clean_word(char *word);
+
+#endif
